requirements.txt
Flask==2.2.3
requests==2.28.2
beautifulsoup4~=4.12.0
mysql-connector-python~=8.0.32